# docs/api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EvolveFit API V1",
    "version": "1.0.0",
    "description": "Local-first EvolveFit API contract for hydration, supplements, routines, workouts, sync, auth, and observability."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "SupabaseBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "SupabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase JWT",
        "description": "Production data endpoints require a Supabase JWT via Authorization: Bearer <token> or an httpOnly session cookie containing the access token."
      }
    },
...
    "schemas": {
...
      "ClientErrorReport": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "digest": {
            "type": "string"
          },
          "stack": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Ok": {
        "description": "Successful API result.",
        "headers": {
          "x-request-id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiOk"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/observability/logs": {
      "get": {
        "summary": "Recent structured logs, optionally filtered by requestId",
        "parameters": [
          {
            "name": "requestId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        },
        "security": []
      }
    },
    "/api/client-errors": {
      "post": {
        "summary": "Report client error",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientErrorReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
    "/api/coach/recommend": {
      "post": {
        "summary": "Generate guarded coach recommendation. Rule insight runs first; AI appears only with sufficient data and configured provider.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/coach/recommendations/{id}/feedback": {
      "post": {
        "summary": "Accept or reject a coach recommendation with optional feedback.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "decision"
                ],
                "properties": {
                  "decision": {
                    "enum": [
                      "accepted",
                      "rejected"
                    ]
                  },
                  "feedback": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/auth/sign-in": {
      "post": {
        "summary": "Sign in",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
    "/api/auth/sign-up": {
      "post": {
        "summary": "Sign up",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
...
    "/api/auth/oauth/google": {
      "get": {
        "summary": "Start Supabase Google OAuth PKCE flow.",
        "responses": {
          "302": {
            "description": "Redirect to provider authorization URL."
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
    "/api/auth/callback": {
      "get": {
        "summary": "Complete Supabase OAuth callback using PKCE code or token query parameters.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "access_token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh_token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires_at",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the app after setting session cookies."
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
...
    "/api/hydration/log": {
      "post": {
        "summary": "Create hydration log",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/hydration/log/{id}": {
      "patch": {
        "summary": "Update hydration log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete hydration log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/supplements": {
      "get": {
        "summary": "List supplements",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      },
      "post": {
        "summary": "Create supplement",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/supplements/log": {
      "post": {
        "summary": "Create supplement log",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/supplements/{id}/reminder": {
      "patch": {
        "summary": "Update a supplement reminder hour.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      },
      "put": {
        "summary": "Update supplement details and schedule fields.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
    "/api/routines": {
      "get": {
        "summary": "List routines",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      },
      "post": {
        "summary": "Create routine",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/routines/{id}": {
      "patch": {
        "summary": "Update routine, may return conflict preview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete routine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/exercises": {
      "get": {
        "summary": "List exercise definitions",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      },
      "post": {
        "summary": "Create exercise definition",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/exercises/{id}": {
      "patch": {
        "summary": "Update exercise definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete exercise definition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/workouts/sessions": {
      "post": {
        "summary": "Start workout session",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/workouts/sessions/{id}/finish": {
      "post": {
        "summary": "Finish workout session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  }
}